// lib/api/model/moduleAttendanceSummary.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ModuleAttendanceSummary {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  moduleId: number | string;
  moduleTitle: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  orderIndex: number | string;
  /** @nullable */
  attendanceStatus: string | null;
  /** @nullable */
  sectionName: string | null;
}